zebra-label-printer/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zebra-label-parser"
version = "0.1.0"
description = "Shipping label extraction and Zebra printer service"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "PyMuPDF>=1.24.0",
    "Pillow>=10.4.0",
    "anthropic>=0.39.0",
    "pydantic-settings>=2.6.0",
]

[project.optional-dependencies]
cups = ["pycups>=2.0.4"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
